requirements.txt
fastapi
uvicorn
firebase-admin
torch
torchvision
Pillow
matplotlib
python-multipart

